requirements.txt
Django
djangorestframework
djangorestframework-api-key
python-dotenv
python-decouple
whitenoise
pymysql
